requirements.txt
pygame>=2.5.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
customtkinter>=5.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pymunk>=7.0.0
